chore: Extract XAML gen error handling in dedicated file
